# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TM Seasons API Docs",
        "version": "1.0.0",
        "description": "REST API for TM Seasons, covering Discord authentication, Trackmania club and map administration, season lifecycle management, leaderboard polling, event-based scoring, standings, and activity feeds."
    },
    "servers": [
        {
            "url": "https://tm-bot.test/api",
            "description": "Local"
        }
    ],
...
    "security": [
        {
            "sanctum": []
        }
    ],
    "paths": {
...
        "/clubs/{club}/members": {
            "get": {
                "operationId": "clubs.members",
                "description": "Returns paginated members for a club, optionally filtered by active membership.",
                "summary": "List public club members",
                "tags": [
                    "Trackmania - Public"
                ],
                "parameters": [
                    {
                        "name": "club",
                        "in": "path",
                        "required": true,
                        "description": "The club ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `ClubMemberResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClubMemberResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                },
                "security": []
            }
        },
...
        "/admin/club": {
            "get": {
                "operationId": "admin.club.primary",
                "description": "Returns the configured primary club, falling back to the oldest imported club when needed.",
                "summary": "Get the primary club",
                "tags": [
                    "Trackmania - Admin"
                ],
                "responses": {
                    "200": {
                        "description": "`TrackmaniaClubResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TrackmaniaClubResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "No primary club configured yet."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/admin/club/sync": {
            "post": {
                "operationId": "admin.club.syncPrimary",
                "description": "Uses the supplied club ID when present, otherwise syncs the configured primary club.",
                "summary": "Sync the primary Trackmania club",
                "tags": [
                    "Trackmania - Admin"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncClubRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "club": {
                                            "$ref": "#/components/schemas/TrackmaniaClubResource"
                                        },
                                        "imported": {},
                                        "updated": {},
                                        "deactivated": {},
                                        "enriched": {
                                            "anyOf": [
                                                {},
                                                {
                                                    "type": "integer",
                                                    "enum": [
                                                        0
                                                    ]
                                                }
                                            ]
                                        },
                                        "total_members": {}
                                    },
                                    "required": [
                                        "club",
                                        "imported",
                                        "updated",
                                        "deactivated",
                                        "enriched",
                                        "total_members"
                                    ]
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Unexpected failure while syncing club."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "No primary club configured yet."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/admin/club/members": {
            "get": {
                "operationId": "admin.club.members",
                "description": "Returns paginated members for the primary club, with optional search, active, and sort filters.",
                "summary": "List primary club members for admins",
                "tags": [
                    "Trackmania - Admin"
                ],
                "responses": {
                    "200": {
                        "description": "Array of `ClubMemberResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClubMemberResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Laravel Sanctum bearer token for protected endpoints.",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ClubMemberResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "trackmania_club_id": {
                        "type": "integer"
                    },
                    "trackmania_player_id": {
                        "type": "integer"
                    },
                    "joined_at": {
                        "type": "string"
                    },
                    "synced_at": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "player": {
                        "$ref": "#/components/schemas/TrackmaniaPlayerResource"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "trackmania_club_id",
                    "trackmania_player_id",
                    "joined_at",
                    "synced_at",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "ClubMemberResource"
            },
...
            "SyncClubRequest": {
                "type": "object",
                "properties": {
                    "club_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    }
                },
                "title": "SyncClubRequest"
            },
            "TrackmaniaClubResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "club_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "tag": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "member_count": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "icon_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_primary": {
                        "type": "boolean"
                    },
                    "last_synced_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "club_id",
                    "name",
                    "tag",
                    "description",
                    "member_count",
                    "icon_url",
                    "is_primary",
                    "last_synced_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "TrackmaniaClubResource"
            },
            "TrackmaniaPlayerResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "account_id": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "zone_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "zone_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "last_seen_in_club_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "account_id",
                    "display_name",
                    "zone_id",
                    "zone_name",
                    "is_active",
                    "last_seen_in_club_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "TrackmaniaPlayerResource"
            },
...
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}